Fix (almost all) compilation warnings

+++ examples/Application/IrrigationSimple/Helpers.h
@@ -11,9 +11,9 @@
 
 void setSystemClock(String date = __DATE__, String time = __TIME__);
 void statusPrint();
-int getAverageInputRead(int pin, size_t loops = 20);
+uint16_t getAverageInputRead(int pin, size_t loops = 20);
 float getAverage05VRead(int pin);
-int getMoisturePerc(int pin);
+uint8_t getMoisturePerc(int pin);
 
 extern std::list<AlarmID_t> alarmTabIDs;
 extern std::list<AlarmID_t> alarmSketchIDs;
