+++ examples/Application/IrrigationSimpleLCD/Helpers.h
@@ -13,9 +13,9 @@
 void setSystemClock(String date = __DATE__, String time = __TIME__);
 void statusLCD();
 void backlightOff(bool powerDown);
-int getAverageInputRead(int pin, size_t loops = 20);
+uint16_t getAverageInputRead(int pin, size_t loops = 20);
 float getAverage05VRead(int pin);
-int getMoisturePerc(int pin);
+uint8_t getMoisturePerc(int pin);
 void displayMsg(const String msg, const unsigned timeout = 5000, const unsigned line = 0, const bool clear = true, const bool off = true);
 
 extern std::list<AlarmID_t> alarmTabIDs;
